Rename package nspmd.util.document to nspmd.lang.document
