+++ pmd-core/src/main/java/net/sourceforge/pmd/PmdAnalysis.java
@@ -21,6 +21,7 @@ import net.sourceforge.pmd.internal.util.FileCollectionUtil;
 import net.sourceforge.pmd.lang.Language;
 import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.LanguageVersionDiscoverer;
+import net.sourceforge.pmd.lang.document.FileCollector;
 import net.sourceforge.pmd.processor.AbstractPMDProcessor;
 import net.sourceforge.pmd.processor.MonoThreadProcessor;
 import net.sourceforge.pmd.processor.MultiThreadProcessor;
@@ -28,7 +29,6 @@ import net.sourceforge.pmd.renderers.Renderer;
 import net.sourceforge.pmd.util.ClasspathClassLoader;
 import net.sourceforge.pmd.util.IOUtil;
 import net.sourceforge.pmd.util.datasource.DataSource;
-import net.sourceforge.pmd.util.document.FileCollector;
 import net.sourceforge.pmd.util.log.PmdLogger;
 import net.sourceforge.pmd.util.log.PmdLogger.Level;
 import net.sourceforge.pmd.util.log.SimplePmdLogger;
