+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/document/TextFile.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.util.document;
+package net.sourceforge.pmd.lang.document;
 
 import java.io.File;
 import java.io.IOException;
